Rename FrameColorBackground and split ColumnRectangle to its own file
